requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
et-xmlfile==1.1.0
fastapi==0.105.0
idna==3.6
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
xarray==2023.12.0
